+++ run.sh
@@ -1,9 +1,0 @@
-
-python3 -m isitfit.cli
-
-isitfit
-
-AWS_DEFAULT_REGION=us-west-2 isitfit
-
-AWS_PROFILE=autofitcloud AWS_DEFAULT_REGION=eu-central-1 isitfit
-