+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version latest (0.5.0?, 2019-09-24?)
 - ...
 - enh: move tests into `isitfit/tests`
 - enh: add dockerfile to test in clean environment
+- enh: move DotMan to separate file (out of utils) with functional tests
 
 
 Version 0.10.1 (2019-10-09)